+++ cypress/fixtures/loginData.json
@@ -1,4 +1,5 @@
 {
   "password": "Test@123",
-  "email": "sahu.shubham3591@gmail.com"
+  "email": "sahu.shubham3591@gmail.com",
+  "password2": "DSVFSDIB"
 }
